requirements.txt
requests
pandas
tqdm
openpyxl
python-dotenv